requirements.txt
Django==4.2.27
Pillow==11.0.0
assemblyai==0.48.1
httpx==0.28.1
websockets==15.0.1
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
soundfile==0.12.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
